refactor: create components folder for EA about
